patch connection class to transfer packages of maximum size

+++ connection.h
@@ -28,7 +28,7 @@ public slots:
 private slots:
 
 private:
-    quint16 blockSize;
+    qint64 blockSize;
     static const QDataStream::Version streamVersion;
     
 };
